step 10

+++ app/src/main/java/hexlet/code/service/LabelService.java
@@ -7,11 +7,11 @@ import hexlet.code.exception.ResourceNotFoundException;
 import hexlet.code.mapper.LabelMapper;
 import hexlet.code.model.Label;
 import hexlet.code.repository.LabelRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class LabelService {
@@ -21,46 +21,40 @@ public class LabelService {
     @Autowired
     private LabelMapper labelMapper;
 
-    public LabelDTO getLabelById(Long id) {
+    public List<LabelDTO> getAll() {
+        return labelRepository.findAllDto();
+    }
+
+    public LabelDTO findById(Long id) {
         Label label = labelRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Label '" + id + "' not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("Label with id '%d' not found", id));
         return labelMapper.map(label);
     }
 
     public LabelDTO findByName(String name) {
-        Optional<Label> labelOptional = labelRepository.findByName(name);
-        if (labelOptional.isEmpty()) {
-            throw new ResourceNotFoundException("Label with name '" + name + "' not found");
-        }
-        return labelMapper.map(labelOptional.get());
-    }
-
-    public List<LabelDTO> getAll() {
-        List<Label> labelList = labelRepository.findAll();
-        return labelList.stream()
-                .map(labelMapper::map)
-                .toList();
-    }
-
-    public LabelDTO create(LabelCreateDTO labelCreateDTO) {
-        Optional<Label> labelOptional = labelRepository.findByName(labelCreateDTO.getName());
-        if (labelOptional.isPresent()) {
-            throw new ResourceNotFoundException("Label with name " + labelCreateDTO.getName() + " already exists");
-        }
-        Label label = labelMapper.map(labelCreateDTO);
-        label = labelRepository.save(label);
+        Label label = labelRepository.findByName(name)
+                .orElseThrow(() -> new ResourceNotFoundException("Label with id '%d' not found", name));
         return labelMapper.map(label);
     }
 
+    @Transactional
+    public LabelDTO create(LabelCreateDTO labelCreateDTO) {
+        return labelMapper.map(labelRepository.save(labelMapper.map(labelCreateDTO)));
+    }
+
+    @Transactional
     public LabelDTO update(Long id, LabelUpdateDTO labelUpdateDTO) {
         Label label = labelRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Label " + id + " not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("Label with id '%d' not found", id));
         labelMapper.update(labelUpdateDTO, label);
-        label = labelRepository.save(label);
-        return labelMapper.map(label);
+        return labelMapper.map(labelRepository.save(label));
     }
 
-    public void delete(Long id) {
+    @Transactional
+    public void deleteById(Long id) {
+        if (!labelRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Label with id '%d' not found", id);
+        }
         labelRepository.deleteById(id);
     }
 }
