+++ app/src/main/java/hexlet/code/service/TaskService.java
@@ -9,6 +9,8 @@ import hexlet.code.mapper.TaskMapper;
 import hexlet.code.model.Task;
 import hexlet.code.repository.TaskRepository;
 import hexlet.code.specification.TaskSpecification;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
@@ -17,13 +19,17 @@ import java.util.List;
 
 @Service
 public class TaskService {
+    private static final Logger LOGGER = LoggerFactory.getLogger(TaskService.class);
+    private final TaskMapper taskMapper;
+    private final TaskRepository taskRepository;
+    private final TaskSpecification taskSpecification;
 
     @Autowired
-    private TaskMapper taskMapper;
-    @Autowired
-    private TaskRepository taskRepository;
-    @Autowired
-    private TaskSpecification taskSpecification;
+    public TaskService(TaskMapper taskMapper, TaskRepository taskRepository, TaskSpecification taskSpecification) {
+        this.taskMapper = taskMapper;
+        this.taskRepository = taskRepository;
+        this.taskSpecification = taskSpecification;
+    }
 
     public List<TaskDTO> getAll(TaskParamsDTO taskParamsDTO) {
         Specification<Task> spec = taskSpecification.build(taskParamsDTO);
@@ -31,28 +37,30 @@ public class TaskService {
     }
 
     public TaskDTO getTaskById(Long id) {
-        Task task = taskRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException(
-                        "Task '" + id + "' not found"));
+        Task task = taskRepository.findById(id).orElseThrow(() ->
+                new ResourceNotFoundException("Task with id '%d' not found", id));
         return taskMapper.map(task);
     }
 
     public TaskDTO create(TaskCreateDTO taskCreateDTO) {
         Task task = taskMapper.map(taskCreateDTO);
         task = taskRepository.save(task);
+        LOGGER.info("Task created with id: {}", task.getId());
         return taskMapper.map(task);
     }
 
     public TaskDTO update(Long id, TaskUpdateDTO taskUpdateDTO) {
-        Task task = taskRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException(
-                        "Task with id '" + id + "' not found"));
+        Task task = taskRepository.findById(id).orElseThrow(() ->
+                new ResourceNotFoundException("Task with id '%d' not found", id));
         taskMapper.update(taskUpdateDTO, task);
         task = taskRepository.save(task);
         return taskMapper.map(task);
     }
 
-    public void delete(Long id) {
+    public void deleteById(Long id) {
+        if (!taskRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Task with id '%d' not found", id);
+        }
         taskRepository.deleteById(id);
     }
 }
