+++ app/src/main/java/hexlet/code/service/UserService.java
@@ -7,21 +7,26 @@ import hexlet.code.exception.ResourceNotFoundException;
 import hexlet.code.mapper.UserMapper;
 import hexlet.code.model.User;
 import hexlet.code.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import jakarta.transaction.Transactional;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
 public class UserService {
+    private static final Logger LOGGER = LoggerFactory.getLogger(UserService.class);
 
-    @Autowired
-    private UserRepository userRepository;
-    @Autowired
-    private UserMapper userMapper;
+    private final UserRepository userRepository;
+    private final UserMapper userMapper;
+
+    public UserService(UserRepository userRepository, UserMapper userMapper) {
+        this.userRepository = userRepository;
+        this.userMapper = userMapper;
+    }
 
     public List<UserDTO> getAllUsers() {
-
         return userRepository.findAll().stream()
                 .map(userMapper::map)
                 .toList();
@@ -29,28 +34,33 @@ public class UserService {
 
     public UserDTO getUserById(Long id) {
         User user = userRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("User " + id + " not found"));
-
+                .orElseThrow(() -> new ResourceNotFoundException("User with id '%d' not found", id));
         return userMapper.map(user);
     }
 
+    @Transactional
     public UserDTO createUser(UserCreateDTO userCreateDTO) {
         User user = userMapper.map(userCreateDTO);
         user = userRepository.save(user);
-
+        LOGGER.info("Created user with id: {}", user.getId());
         return userMapper.map(user);
     }
 
+    @Transactional
     public UserDTO updateUser(Long id, UserUpdateDTO userUpdateDTO) {
-        User user = userRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("User '" + id + "' not found"));
+        User user = userRepository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("User with id '%d' not found", id));
         userMapper.updateUser(userUpdateDTO, user);
-        user = userRepository.save(user);
-
+        LOGGER.info("Updated user with id: {}", id);
         return userMapper.map(user);
     }
 
-    public void deleteUser(Long id) {
+    @Transactional
+    public void deleteById(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new ResourceNotFoundException("User with id '%d' not found", id);
+        }
         userRepository.deleteById(id);
+        LOGGER.info("Deleted user with id: {}", id);
     }
 }
